Add detached signature API
For liboqs integration, in addition to the standard keypair, sign, open API,
additional functions creating and verifying standalone signatures are required.

This commit adds these functions and implements the existing sign and open
functions on top of them.

This also resolves a bug in the previous code that would not correclty
zero out the message buffer in case the signature verification fails.