OpenSSL: EVP_MD_CTX_{create,destroy} -> {new, free}
The EVP_MD_CTX_create() and EVP_MD_CTX_destroy() functions were renamed to
EVP_MD_CTX_new() and EVP_MD_CTX_free() in OpenSSL 1.1.0, respectively.
The old names still appear to be supported, but are deprecated.
We should change to the new names to not risk breaking with future OpenSSL
versions.